♻️ Refactor: 스타일 파일, 타입 파일 파일명 변경 및 적용 (#23)
