+++ src/components/common/input/PasswordInput.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import DefaultInput from './DefaultInput';
 import { EyeIcon, EyeSlashIcon } from '@heroicons/react/24/outline';
-import { iconStyle } from './InputStyles';
-import type { InputProps } from './types';
+import { iconStyle } from './input.styles';
+import type { InputProps } from './input.types';
 
 function PasswordInput({ label }: InputProps) {
   const [isHidden, setIsHidden] = useState<boolean>(true);
