+++ src/components/common/input/TimeInput.tsx
@@ -1,8 +1,8 @@
 import { useRef, useState } from 'react';
 import DefaultInput from './DefaultInput';
 import { ClockIcon } from '@heroicons/react/24/outline';
-import { iconStyle, timeStyle } from './InputStyles';
-import type { InputProps } from './types';
+import { iconStyle, timeStyle } from './input.styles';
+import type { InputProps } from './input.types';
 import { useClickOutside } from '@/hooks/useClickOutside';
 
 function TimeInput({ label }: InputProps) {
